digicloset-upgrade-pack/model-service/requirements.txt
fastapi==0.95.2
uvicorn==0.22.0
torch>=2.0.0
pillow==10.1.0
aiofiles==23.1.0
python-multipart==0.0.6
diffusers>=0.24.0
transformers>=4.36.0
accelerate>=0.25.0
opencv-python-headless>=4.8.0
scikit-image>=0.21.0
httpx>=0.26.0
numpy>=1.26.0,<2.0.0
paramiko>=2.7.2
types-paramiko>=3.0.0
controlnet-aux>=0.0.7
mediapipe>=0.10.9
